Refactored line22.
